+++ macros/grants/grant_privileges.sql
@@ -1,36 +1,40 @@
-{% macro grant_privileges(schemas) %}
+{% macro grant_privileges(domain_schemas) %}
     {% if flags.WHICH == 'run' %}
-        {%- set domain_schemas = ["SALES", "DM", "CUSTOMER"] -%}
         {% if target.name == 'prod' %}
-            {{ dbt_dataengineers_utils.grant_database_ownership_access('DATAOPS_ADMIN') }}
-            {{ dbt_dataengineers_utils.grant_database_usage_access('DATAOPS_ADMIN') }}
-            {{ dbt_dataengineers_utils.grant_database_usage_access('READERS_PROD') }}
-            {{ dbt_dataengineers_utils.grant_database_usage_access('ANALYST') }}
-            {{ dbt_dataengineers_utils.grant_schema_ownership_access(schemas, 'DATAOPS_ADMIN', false) }}
-            {{ dbt_dataengineers_utils.grant_schema_read_access(schemas, 'ANALYST', false) }}
-            {{ dbt_dataengineers_utils.grant_schema_read_access(domain_schemas, 'READERS_PROD', false) }}
-        {% elif target.name == 'unit-test' %}
-            {{ dbt_dataengineers_utils.grant_database_ownership_access('DATAOPS_ADMIN') }}
-            {{ dbt_dataengineers_utils.grant_database_usage_access('DATAOPS_ADMIN') }}
-            {{ dbt_dataengineers_utils.grant_database_usage_access('DEVELOPERS') }}
-            {{ dbt_dataengineers_utils.grant_schema_ownership_access(schemas, 'DATAOPS_ADMIN', false) }}
-            {{ dbt_dataengineers_utils.grant_schema_read_access(schemas, 'DEVELOPERS', false) }}
+            {{ dbt_dataengineers_utils.grant_database_ownership('DATAOPS_ADMIN') }}
+            {{ dbt_dataengineers_utils.grant_database_usage(['READERS_PROD', 'ANALYST', 'OPS_SUPPORT']) }}
+            {{ dbt_dataengineers_utils.grant_schema_ownership([], 'DATAOPS_ADMIN') }}
+            {{ dbt_dataengineers_utils.grant_schema_operate([], ['OPS_SUPPORT']) }}
+            {{ dbt_dataengineers_utils.grant_schema_monitor([], ['OPS_SUPPORT']) }}
+            {{ dbt_dataengineers_utils.grant_schema_read([], ['ANALYST'], false) }}
+            {{ dbt_dataengineers_utils.grant_schema_read_specific(domain_schemas, ['READERS_PROD'], false, false) }}
         {% elif target.name == 'test' %}
-            {{ dbt_dataengineers_utils.grant_database_ownership_access('DATAOPS_ADMIN') }}
-            {{ dbt_dataengineers_utils.grant_database_usage_access('DATAOPS_ADMIN') }}
-            {{ dbt_dataengineers_utils.grant_database_usage_access('READERS_TEST') }}
-            {{ dbt_dataengineers_utils.grant_database_usage_access('ANALYST') }}
-            {{ dbt_dataengineers_utils.grant_schema_ownership_access(schemas, 'DATAOPS_ADMIN', false) }}
-            {{ dbt_dataengineers_utils.grant_schema_read_access(schemas, 'ANALYST', false) }}
-            {{ dbt_dataengineers_utils.grant_schema_read_access(domain_schemas, 'READERS_TEST', false) }}
+            {{ dbt_dataengineers_utils.grant_database_ownership('DATAOPS_ADMIN') }}
+            {{ dbt_dataengineers_utils.grant_database_usage(['READERS_TEST', 'ANALYST', "OPS_SUPPORT"]) }}
+            {{ dbt_dataengineers_utils.grant_schema_ownership([], 'DATAOPS_ADMIN') }}
+            {{ dbt_dataengineers_utils.grant_schema_operate([], ['OPS_SUPPORT']) }}
+            {{ dbt_dataengineers_utils.grant_schema_monitor([], ['OPS_SUPPORT']) }}
+            {{ dbt_dataengineers_utils.grant_schema_read([], ['ANALYST'], false) }}
+            {{ dbt_dataengineers_utils.grant_schema_read_specific(domain_schemas, ['READERS_TEST'], false, false) }}
+        {% elif target.name == 'unit-test' %}
+            {{ dbt_dataengineers_utils.grant_database_ownership('DATAOPS_ADMIN') }}
+            {{ dbt_dataengineers_utils.grant_database_usage(['DEVELOPERS']) }}
+            {{ dbt_dataengineers_utils.grant_schema_ownership([], 'DATAOPS_ADMIN') }}
+            {{ dbt_dataengineers_utils.grant_schema_read([], ['DEVELOPERS'], false) }}
         {% elif target.name == 'dev' %}
-            {{ dbt_dataengineers_utils.grant_database_ownership_access('DEVELOPERS') }}
-            {{ dbt_dataengineers_utils.grant_database_usage_access('DEVELOPERS') }}
-            {{ dbt_dataengineers_utils.grant_schema_ownership_access(domain_schemas, 'DEVELOPERS', false) }}
+            {{ dbt_dataengineers_utils.grant_database_ownership('DEVELOPERS') }}
+            {{ dbt_dataengineers_utils.grant_database_usage([]) }}
+            {{ dbt_dataengineers_utils.grant_schema_ownership([], 'DEVELOPERS') }}
+            {{ dbt_dataengineers_utils.grant_schema_monitor([], []) }}
+            {{ dbt_dataengineers_utils.grant_schema_operate([], ['OPS_SUPPORT'], false) }}
+            {{ dbt_dataengineers_utils.grant_schema_read([], [], false) }}
         {% else %}
-            {{ dbt_dataengineers_utils.grant_database_ownership_access('DEVELOPERS') }}
-            {{ dbt_dataengineers_utils.grant_database_usage_access('DEVELOPERS') }}
-            {{ dbt_dataengineers_utils.grant_schema_ownership_access(schemas, 'DEVELOPERS', false) }}
+            {{ dbt_dataengineers_utils.grant_database_ownership('DEVELOPERS') }}
+            {{ dbt_dataengineers_utils.grant_schema_ownership([], 'DEVELOPERS') }}
+            {{ dbt_dataengineers_utils.grant_database_usage([]) }}
+            {{ dbt_dataengineers_utils.grant_schema_monitor([], []) }}
+            {{ dbt_dataengineers_utils.grant_schema_operate([], []) }}
+            {{ dbt_dataengineers_utils.grant_schema_read([], [], false) }}
         {% endif %}
     {% endif %}
 {% endmacro %}